insertion of user restriction

+++ routes/tutorial/deleteTutorial.js
@@ -14,29 +14,37 @@ const deleteTutorial = async function(req, res){
   try{
     var tutorial = await Tutorial.findById(req.params.tutorialId);
     if(tutorial){
-      await Tutorial.deleteOne({_id: req.params.tutorialId});
-      // // remove links to the deleted tutorial
-      await Tutorial.updateMany({}, {$pull: {'steps.$[].requirements': req.params.tutorialId}}, {multi: true});
-      // remove images of the deleted tutorial
-      var imagePaths = [];
-      tutorial.steps.forEach((step, i) => {
-        if(step.media && step.media.picture && step.media.picture.path){
-          imagePaths.push(step.media.picture.path);
-        }
-      });
-      imagePaths.forEach((imagePath, i) => {
-        fs.unlink(path.join(__dirname, '..', '..', 'upload', imagePath), function(err) {
-          // if(err && err.code == 'ENOENT') {
-            // file doens't exist
-          // } else if (err) {
-            // other errors, e.g. maybe we don't have enough permission
-          // } else {
-          // }
+      var owner = req.user.email;
+      if(owner === tutorial.creator){
+        await Tutorial.deleteOne({_id: req.params.tutorialId});
+        // // remove links to the deleted tutorial
+        await Tutorial.updateMany({}, {$pull: {'steps.$[].requirements': req.params.tutorialId}}, {multi: true});
+        // remove images of the deleted tutorial
+        var imagePaths = [];
+        tutorial.steps.forEach((step, i) => {
+          if(step.media && step.media.picture && step.media.picture.path){
+            imagePaths.push(step.media.picture.path);
+          }
         });
-      });
-      return res.status(200).send({
-        message: 'Tutorial deleted successfully.',
-      });
+        imagePaths.forEach((imagePath, i) => {
+          fs.unlink(path.join(__dirname, '..', '..', 'upload', imagePath), function(err) {
+            // if(err && err.code == 'ENOENT') {
+              // file doens't exist
+            // } else if (err) {
+              // other errors, e.g. maybe we don't have enough permission
+            // } else {
+            // }
+          });
+        });
+        return res.status(200).send({
+          message: 'Tutorial deleted successfully.',
+        });
+      }
+      else {
+        return res.status(403).send({
+          message: 'No permission deleting the tutorial.',
+        });
+      }
     }
     return res.status(404).send({
       message: 'Tutorial not found.',
